test: rename e2e import aliases off the fetch prefix
The PR removes the fetch prefix from the list API, but the e2e import
aliases kept it (fetchProjects, fetchIssues), reintroducing the naming
seam locally. Applied across every e2e file rather than only the three
flagged ones so the aliases stay uniform.

Addresses the gemini-code-assist review comments on #444.
