+++ e2e/journals.test.ts
@@ -5,7 +5,7 @@ import { show } from "../issues/show.ts";
 import { createIssue } from "../issues/create.ts";
 import { update } from "../issues/update.ts";
 import { deleteIssue } from "../issues/delete.ts";
-import { list as fetchProjects } from "../projects/list.ts";
+import { list as listProjects } from "../projects/list.ts";
 import { update as updateJournal } from "../journals/update.ts";
 
 Deno.test({
@@ -18,7 +18,7 @@ Deno.test({
     await t.step(
       "POST /issues.json should create an issue to attach a journal to",
       async () => {
-        const projects = await fetchProjects(e2eContext);
+        const projects = await listProjects(e2eContext);
         const project = projects.find((p) =>
           p.identifier === "e2e-test-project"
         );
