+++ e2e/time-entries.test.ts
@@ -5,17 +5,17 @@ import { show } from "../time-entries/show.ts";
 import { create } from "../time-entries/create.ts";
 import { update } from "../time-entries/update.ts";
 import { deleteTimeEntry } from "../time-entries/delete.ts";
-import { list as fetchProjects } from "../projects/list.ts";
-import { list as fetchIssues } from "../issues/list.ts";
+import { list as listProjects } from "../projects/list.ts";
+import { list as listIssues } from "../issues/list.ts";
 
 Deno.test({
   name: "E2E: Time Entries API",
   fn: async (t) => {
-    const projects = await fetchProjects(e2eContext);
+    const projects = await listProjects(e2eContext);
     const project = projects.find((p) => p.identifier === "e2e-test-project");
     expect(project).toBeDefined();
 
-    const issues = await fetchIssues(e2eContext, {
+    const issues = await listIssues(e2eContext, {
       projectId: project!.id,
     });
     expect(issues.length).toBeGreaterThan(0);
